feat: update grocery classes

+++ src/main/java/no/ntnu/idatt1005/Grocery/Grocery.java
@@ -6,12 +6,12 @@ import no.ntnu.idatt1005.Unit.UnitsE;
 public class Grocery {
     private String name;
     private String image;
-    private float amount;
+    private double amount;
     private final String id;
     private UnitsE unit;
 
 
-    Grocery(String name, String image, float amount, String id, UnitsE unit) {
+    Grocery(String name, String image, double amount, String id, UnitsE unit) {
         this.name = name;
         this.image = image;
         this.amount = amount;
@@ -26,7 +26,7 @@ public class Grocery {
         return image;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 
@@ -38,7 +38,7 @@ public class Grocery {
         return unit;
     }
 
-    public float setAmount(float input) {
+    public double setAmount(double input) {
         return amount;
     }
